# frontend/openapi.json
{
  "components" : {
    "schemas" : {
...
      "MonthSummary" : {
        "properties" : {
          "inflow" : {
            "type" : "string"
          },
          "month" : {
            "type" : "string"
          },
          "outflow" : {
            "type" : "string"
          }
        },
        "required" : [ "inflow", "month", "outflow" ],
        "type" : "object"
      },
...
      "PageTransactionResponse" : {
        "properties" : {
          "content" : {
            "items" : {
              "$ref" : "#/components/schemas/TransactionResponse"
            },
            "type" : "array"
          },
          "empty" : {
            "type" : "boolean"
          },
          "first" : {
            "type" : "boolean"
          },
          "last" : {
            "type" : "boolean"
          },
          "number" : {
            "format" : "int32",
            "type" : "integer"
          },
          "numberOfElements" : {
            "format" : "int32",
            "type" : "integer"
          },
          "pageable" : {
            "$ref" : "#/components/schemas/PageableObject"
          },
          "size" : {
            "format" : "int32",
            "type" : "integer"
          },
          "sort" : {
            "$ref" : "#/components/schemas/SortObject"
          },
          "totalElements" : {
            "format" : "int64",
            "type" : "integer"
          },
          "totalPages" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "required" : [ "content", "empty", "first", "last", "number", "numberOfElements", "pageable", "size", "sort", "totalElements", "totalPages" ],
        "type" : "object"
      },
...
      "Pageable" : {
        "properties" : {
          "page" : {
            "format" : "int32",
            "minimum" : 0,
            "type" : "integer"
          },
          "size" : {
            "format" : "int32",
            "minimum" : 1,
            "type" : "integer"
          },
          "sort" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          }
        },
        "type" : "object"
      },
      "PageableObject" : {
        "properties" : {
          "offset" : {
            "format" : "int64",
            "type" : "integer"
          },
          "pageNumber" : {
            "format" : "int32",
            "type" : "integer"
          },
          "pageSize" : {
            "format" : "int32",
            "type" : "integer"
          },
          "paged" : {
            "type" : "boolean"
          },
          "sort" : {
            "$ref" : "#/components/schemas/SortObject"
          },
          "unpaged" : {
            "type" : "boolean"
          }
        },
        "required" : [ "offset", "pageNumber", "pageSize", "paged", "sort", "unpaged" ],
        "type" : "object"
      },
...
      "SortObject" : {
        "properties" : {
          "empty" : {
            "type" : "boolean"
          },
          "sorted" : {
            "type" : "boolean"
          },
          "unsorted" : {
            "type" : "boolean"
          }
        },
        "required" : [ "empty", "sorted", "unsorted" ],
        "type" : "object"
      },
...
      "TransactionHistoryPage" : {
        "properties" : {
          "items" : {
            "items" : {
              "$ref" : "#/components/schemas/TransactionResponse"
            },
            "type" : "array"
          },
          "nextCursor" : {
            "type" : "string"
          },
          "total" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "required" : [ "items", "nextCursor", "total" ],
        "type" : "object"
      },
      "TransactionResponse" : {
        "properties" : {
          "amount" : {
            "type" : "string"
          },
          "createdAt" : {
            "type" : "string"
          },
          "currency" : {
            "type" : "string"
          },
          "flagged" : {
            "type" : "boolean"
          },
          "fromIban" : {
            "description" : "null on deposits/credits (no originator)",
            "type" : [ "string", "null" ]
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "kind" : {
            "enum" : [ "TRANSFER", "DEPOSIT", "INTEREST" ],
            "type" : "string"
          },
          "memo" : {
            "type" : [ "string", "null" ]
          },
          "postedAt" : {
            "description" : "null while HELD or when CANCELLED - money has not moved",
            "type" : [ "string", "null" ]
          },
          "reviewed" : {
            "type" : "boolean"
          },
          "status" : {
            "enum" : [ "POSTED", "HELD", "CANCELLED" ],
            "type" : "string"
          },
          "toIban" : {
            "description" : "null on engine loan charges (no destination)",
            "type" : [ "string", "null" ]
          }
        },
        "required" : [ "amount", "createdAt", "currency", "flagged", "fromIban", "id", "kind", "memo", "postedAt", "reviewed", "status", "toIban" ],
        "type" : "object"
      },
      "TransferRequest" : {
        "properties" : {
          "amount" : {
            "pattern" : "^\\d+(\\.\\d{1,4})?$",
            "type" : "string"
          },
          "currency" : {
            "pattern" : "(?i)\\s*USD\\s*",
            "type" : "string"
          },
          "fromAccountId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "memo" : {
            "maxLength" : 140,
            "minLength" : 0,
            "type" : "string"
          },
          "toIban" : {
            "minLength" : 1,
            "type" : "string"
          }
        },
        "required" : [ "amount", "toIban" ],
        "type" : "object"
      },
      "TransferResponse" : {
        "properties" : {
          "amount" : {
            "type" : "string"
          },
          "createdAt" : {
            "type" : "string"
          },
          "currency" : {
            "type" : "string"
          },
          "flagged" : {
            "type" : "boolean"
          },
          "fromIban" : {
            "description" : "null on deposits (no originator)",
            "type" : [ "string", "null" ]
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "memo" : {
            "type" : [ "string", "null" ]
          },
          "postedAt" : {
            "description" : "null while HELD or when CANCELLED - money has not moved",
            "type" : [ "string", "null" ]
          },
          "status" : {
            "enum" : [ "POSTED", "HELD", "CANCELLED" ],
            "type" : "string"
          },
          "toIban" : {
            "type" : [ "string", "null" ]
          }
        },
        "required" : [ "amount", "createdAt", "currency", "flagged", "fromIban", "id", "memo", "postedAt", "status", "toIban" ],
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "bearer" : {
        "bearerFormat" : "JWT",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
  "info" : {
    "description" : "Full-stack online banking demo. Errors are RFC-7807 (ApiProblem); money travels as JSON strings.",
    "title" : "Online Banking Simulator API",
    "version" : "1.3.0"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/v1/accounts/{id}/statement.csv" : {
      "get" : {
        "operationId" : "statement",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "from",
          "required" : false,
          "schema" : {
            "format" : "date",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "to",
          "required" : false,
          "schema" : {
            "format" : "date",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "string"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "transfer-controller" ]
      }
    },
    "/api/v1/accounts/{id}/statement.pdf" : {
      "get" : {
        "operationId" : "statementPdf",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "from",
          "required" : false,
          "schema" : {
            "format" : "date",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "to",
          "required" : false,
          "schema" : {
            "format" : "date",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "format" : "byte",
                  "type" : "string"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "transfer-controller" ]
      }
    },
    "/api/v1/accounts/{id}/summary" : {
      "get" : {
        "operationId" : "summary",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "months",
          "required" : false,
          "schema" : {
            "default" : 6,
            "format" : "int32",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/MonthSummary"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "transfer-controller" ]
      }
    },
...
    "/api/v1/admin/transactions" : {
      "get" : {
        "operationId" : "transactions",
        "parameters" : [ {
          "in" : "query",
          "name" : "accountId",
          "required" : false,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "flagged",
          "required" : false,
          "schema" : {
            "type" : "boolean"
          }
        }, {
          "in" : "query",
          "name" : "reviewed",
          "required" : false,
          "schema" : {
            "type" : "boolean"
          }
        }, {
          "in" : "query",
          "name" : "from",
          "required" : false,
          "schema" : {
            "format" : "date",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "to",
          "required" : false,
          "schema" : {
            "format" : "date",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "pageable",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/Pageable"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageTransactionResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "admin-controller" ]
      }
    },
...
    "/api/v1/operations" : {
      "get" : {
        "operationId" : "operation",
        "parameters" : [ {
          "in" : "query",
          "name" : "key",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TransactionResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "transfer-controller" ]
      }
    },
    "/api/v1/transactions" : {
      "get" : {
        "operationId" : "history",
        "parameters" : [ {
          "in" : "query",
          "name" : "accountId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "from",
          "required" : false,
          "schema" : {
            "format" : "date",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "to",
          "required" : false,
          "schema" : {
            "format" : "date",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "size",
          "required" : false,
          "schema" : {
            "default" : 20,
            "format" : "int32",
            "type" : "integer"
          }
        }, {
          "in" : "query",
          "name" : "cursor",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TransactionHistoryPage"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "transfer-controller" ]
      }
    },
    "/api/v1/transfers" : {
      "post" : {
        "operationId" : "transfer",
        "parameters" : [ {
          "in" : "header",
          "name" : "Idempotency-Key",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/TransferRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TransferResponse"
                }
              }
            },
            "description" : "Created"
          }
        },
        "tags" : [ "transfer-controller" ]
      }
    },
    "/api/v1/transfers/{id}" : {
      "get" : {
        "operationId" : "transferDetail",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TransactionResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "transfer-controller" ]
      }
    }
  },
  "security" : [ {
    "bearer" : [ ]
  } ],
  "servers" : [ {
    "url" : "http://localhost:8080"
  } ]
}